requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
Flask
pyyaml
-e .